Initialize work vectors for triangle drawing before the draw loop is called
This gives a fairly substantial speedup by removing a big reason for the slowdown in the first place. Because the vectors are initialized on the heap, the system is doing a lot of unneccessary work every frame by allocating several vectors and then having GC clean up for every triangle in the scene. By creating our work vectors before triangle computations begin and simply manipulating the existing vectors we can save a lot of time